Created new ``pipeline_classification.py`` file
* Moved code for creating a classification lookup table into this file

* Removed some stale imports
